refactor: refactor logged header with the general store states
