+++ frontend/src/app/(auxiliary-functions)/hooks/useSignIn.ts
@@ -1,16 +1,22 @@
 'use client'
 
 import { axiosInstance } from '@/common/axios-api/axios-api'
-import { useContextWrap } from '@/contextAPI/context'
 import { authStoreInstance } from '@/stores/auth/auth.provider'
+import { useGeneralStore } from '@/stores/general/general.provider'
 import signInSchema from '@/zodSchemas/signInSchema'
 import axios from 'axios'
 import { useRouter } from 'next/navigation'
 import { z } from 'zod'
+import { useShallow } from 'zustand/shallow'
 
 export default function useSignIn() {
   type signInData = z.infer<typeof signInSchema>
-  const { setStatusMessage } = useContextWrap()
+  const { setUsername, setStatusMessage } = useGeneralStore(
+    useShallow((store) => ({
+      setUsername: store.setUsername,
+      setStatusMessage: store.setStatusMessage,
+    })),
+  )
   const router = useRouter()
   const setToken = authStoreInstance.getState().setToken
   async function loginSubmitHandler(usersSign: signInData) {
@@ -18,7 +24,9 @@ export default function useSignIn() {
       const response = await axiosInstance.post('/auth/sign-in', usersSign)
       if (response.status === 200) {
         const userToken = response.data.token as string
+        const userName = response.data.username as string
         setToken(userToken)
+        setUsername(userName)
         router.push('/portal/dashboard')
       }
     } catch (error: unknown) {
