+++ frontend/src/app/portal/dashboard/(hooks)/useDeleteWedding.ts
@@ -1,5 +1,5 @@
 import { AxiosApi } from '@/common/axios-api/axios-api'
-import AxiosErrorHandler from '@/functions/axios-error-handler'
+import AxiosErrorHandler from '@/app/(auxiliary-functions)/axios-error-handler'
 import { useWeddingsStore } from '@/stores/weddings/weddings.provider'
 
 import { useRouter } from 'next/navigation'
