+++ frontend/src/app/portal/giftList/[id]/(hooks)/useDeleteGift.ts
@@ -1,5 +1,5 @@
 import { AxiosApi } from '@/common/axios-api/axios-api'
-import AxiosErrorHandler from '@/functions/axios-error-handler'
+import AxiosErrorHandler from '@/app/(auxiliary-functions)/axios-error-handler'
 import { useGiftsStore } from '@/stores/gifts/gifts.provider'
 import { useRouter } from 'next/navigation'
 import { useShallow } from 'zustand/shallow'
