+++ frontend/src/app/portal/giftList/[id]/(hooks)/useGetGifts.ts
@@ -1,6 +1,6 @@
 'use client'
 import { AxiosApi } from '@/common/axios-api/axios-api'
-import AxiosErrorHandler from '@/functions/axios-error-handler'
+import AxiosErrorHandler from '@/app/(auxiliary-functions)/axios-error-handler'
 import { useGiftsStore } from '@/stores/gifts/gifts.provider'
 import { listHeaderProps, weddingGiftsProps } from '@/stores/gifts/gifts.store'
 import { useParams, useRouter } from 'next/navigation'
