--- frontend/src/functions/useCheckAuth.ts
@@ -1,24 +1,0 @@
-'use client'
-
-import { useContextWrap } from '@/contextAPI/context'
-import { useRouter } from 'next/navigation'
-import { useEffect } from 'react'
-
-export default function useCheckAuth() {
-  const route = useRouter()
-  const { setUserToken } = useContextWrap()
-
-  useEffect(() => {
-    function checkAuth() {
-      const storedToken = localStorage.getItem('userToken')
-      if (storedToken) {
-        setUserToken(JSON.parse(storedToken))
-      }
-
-      if (!storedToken) {
-        route.push('/401-page')
-      }
-    }
-    checkAuth()
-  }, [setUserToken, route])
-}
